+++ tests/Phake/Matchers/AbstractChainableArgumentMatcherTest.php
@@ -79,7 +79,7 @@ class AbstractChainableArgumentMatcherTest extends TestCase
 
     public function testSetNextMatcherCallsAssertPreviousMatcher(): void
     {
-        Phake::when($this->nextMatcher)->assertPreviousMatcher->thenThrow(new \InvalidArgumentException());
+        Phake::when($this->nextMatcher)->assertPreviousMatcher(Phake::anyParameters())->thenThrow(new \InvalidArgumentException());
 
         $this->expectException(\InvalidArgumentException::class);
         $this->matcher->setNextMatcher($this->nextMatcher);